+++ pkg/services/rpcsrv/error.go
@@ -1,4 +1,4 @@
-package server
+package rpcsrv
 
 import (
 	"net/http"
