+++ pkg/services/rpcsrv/prometheus.go
@@ -1,4 +1,4 @@
-package server
+package rpcsrv
 
 import (
 	"fmt"
