+++ pkg/services/rpcsrv/subscription_test.go
@@ -1,4 +1,4 @@
-package server
+package rpcsrv
 
 import (
 	"encoding/json"
