+++ pkg/services/rpcsrv/util.go
@@ -1,4 +1,4 @@
-package server
+package rpcsrv
 
 import (
 	"errors"
